requirements.txt
addict==2.4.0
discord-py-interactions==4.2.1
dnspython==2.1.0
pillow==8.3.2
pymongo==3.12.0
requests==2.25.1